extracted table of contents for docs pages into a separate file to prevent unnecessary rerenders
