+++ app/[product]/docs/[slug]/TableOfContents.tsx
@@ -1,12 +1,10 @@
 // app/[product]/docs/[slug]/TableOfContents.tsx
 import client from "@tina/__generated__/client";
-import { TableOfContentsClient } from "./TableOfContentsClient";
+import TableOfContentsClient from "./TableOfContentsClient";
 
-export default async function TableOfContents({
-  product,
-}: {
-  product: string;
-}) {
+export default async function TableOfContents({}) {
+  // console.log("Fetching Table of Contents for product:", product);
+  const product = "YakShaver";
   // Fetch the data on the server
   const tableOfContentsData = await getDocsTableOfContents(product);
   return (
